requirements.txt
Django==4.1.3
firebase-admin
pyjwt
psycopg2-binary
whitenoise
pytz